+++ UniversityCourseSelection/src/main/java/com/capgemini/UniversityCourseSelection/services/AdmissionCommitteeMemberServiceImpl.java
@@ -1,7 +1,10 @@
 package com.capgemini.UniversityCourseSelection.services;
 
 import java.time.LocalDate;
+import java.time.temporal.ChronoUnit;
 import java.util.List;
+
+import javax.transaction.Transactional;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Component;
@@ -11,6 +14,7 @@ import com.capgemini.UniversityCourseSelection.entities.AdmissionCommiteeMember;
 import com.capgemini.UniversityCourseSelection.entities.AdmissionStatus;
 import com.capgemini.UniversityCourseSelection.entities.Applicant;
 import com.capgemini.UniversityCourseSelection.entities.Course;
+import com.capgemini.UniversityCourseSelection.exception.NotFoundException;
 import com.capgemini.UniversityCourseSelection.repo.IAdmissionCommiteeMemberRepository;
 import com.capgemini.UniversityCourseSelection.repo.IApplicantRepository;
 
@@ -29,19 +33,32 @@ public class AdmissionCommitteeMemberServiceImpl implements IAdmissionCommiteeMe
 	}
 
 	@Override
+	@Transactional
 	public AdmissionCommiteeMember updateCommitteeMember(AdmissionCommiteeMember member) {
-		return repo.save(member);
+		if (repo.existsById(member.getAdminId())) {
+			return repo.save(member);
+		} else {
+			throw new NotFoundException("AdmissionCommitteeMember not available");
+		}
 	}
 
 	@Override
 	public AdmissionCommiteeMember viewCommitteeMember(int id) {
-		AdmissionCommiteeMember member = repo.findById(id).get();
-		return member;
+		if (repo.existsById(id)) {
+			return repo.findById(id).get();
+		} else {
+			throw new NotFoundException("AdmissionCommittee Member not found !");
+		}
 	}
 
 	@Override
+	@Transactional
 	public void removeCommitteeMember(int id) {
-		repo.deleteById(id);
+		if (repo.existsById(id)) {
+			repo.deleteById(id);
+		} else {
+			throw new NotFoundException("No member exists with id:" + id);
+		}
 	}
 
 	@Override
@@ -53,7 +70,7 @@ public class AdmissionCommitteeMemberServiceImpl implements IAdmissionCommiteeMe
 	@Override
 	public AdmissionStatus provideAdmissionResult(Applicant applicant, Admission admission) {
 
-		// get the course object
+//		 get the course object
 		Course course = null;
 		int id = admission.getCourseId();
 		course = repo.getCourseById(id);
@@ -61,6 +78,7 @@ public class AdmissionCommitteeMemberServiceImpl implements IAdmissionCommiteeMe
 		// if course is not found, return admission status as rejected/pending
 		if (course == null) {
 			applicant.setStatus(AdmissionStatus.PENDING);
+			admission.setStatus(AdmissionStatus.PENDING);
 		}
 
 		// criteria 1 (admission date)
@@ -69,29 +87,31 @@ public class AdmissionCommitteeMemberServiceImpl implements IAdmissionCommiteeMe
 
 		if (admissionDate.isAfter(courseStartDate)) {
 			applicant.setStatus(AdmissionStatus.REJECTED);
+			admission.setStatus(AdmissionStatus.REJECTED);
+			
+			applicantRepo.save(applicant);
+			return applicant.getStatus();
 		}
-		
 		// criteria 1 satisfied
-
+		
+		
 		// criteria 2 ( percentage )
 
 		double courseCriteria = course.getCourseCriteria();
-
-
 		double marks = applicant.getApplicantGraduationPercentage();
 
 		if (marks < courseCriteria) {
 			applicant.setStatus(AdmissionStatus.PENDING);
-		}
-		else
-		{
+			admission.setStatus(AdmissionStatus.PENDING);
+		} else {
+			admission.setStatus(AdmissionStatus.CONFIRMED);
 			applicant.setStatus(AdmissionStatus.CONFIRMED);
 		}
 
 		// criteria 2 satisfied
 		applicantRepo.save(applicant);
 		return applicant.getStatus();
-		
+
 	}
 
 }
